Move plugin types to provider types
