+++ packages/x-charts/src/ChartContainer/useChartContainerHooks.ts
@@ -2,8 +2,8 @@ import useForkRef from '@mui/utils/useForkRef';
 import * as React from 'react';
 import { usePluginsMerge } from './usePluginsMerge';
 import { useReducedMotion } from '../hooks/useReducedMotion';
-import { ChartsPluginType } from '../models';
 import { ChartSeriesType } from '../models/seriesType/config';
+import { ChartsPluginType } from '../context/PluginProvider';
 
 export const useChartContainerHooks = (
   ref: React.ForwardedRef<unknown> | null,
