+++ packages/x-charts/src/ScatterChart/legend.ts
@@ -1,6 +1,6 @@
 import { LegendItemParams } from '../ChartsLegend/chartsLegend.types';
 import { getLabel } from '../internals/getLabel';
-import { LegendGetter } from '../models/seriesType/config';
+import { LegendGetter } from '../context/PluginProvider';
 
 const legendGetter: LegendGetter<'scatter'> = (params) => {
   const { seriesOrder, series } = params;
